+++ src/store/action.js
@@ -9,7 +9,6 @@ export const ActionType = {
   GET_FILM_ID: `GET_FILM_ID`,
   GET_COMMENTS: `GET_COMMENTS`,
   GET_PROMO_FILM: `GET_PROMO_FILM`,
-  LOADING: `LOADING`,
   RESET_FILM: `RESET_FILM`,
   NOT_FOUND: `NOT_FOUND`,
 };
@@ -53,9 +52,6 @@ export const ActionCreator = {
   loadComments: (comments) => ({
     type: ActionType.GET_COMMENTS,
     payload: comments
-  }),
-  loading: () => ({
-    type: ActionType.LOADING
   }),
   resetFilm: () => ({
     type: ActionType.RESET_FILM
